define, error 한 헤더에 몰아서 저장
