+++ .gitignore
@@ -1,4 +1,2 @@
 a.out
 webserv
-/include/test.hpp
-test.cpp
