+++ default_conf.hpp
@@ -4,4 +4,11 @@
 #define WEBSERV_CONF_PATH "./webserv.conf"
 #define DEFAULT_HTML_PATH "./html_file"
 
+#define BUF_SIZE 			100
+#define	RQ_LINE_NOT_PARSED	0
+#define	HEADER_NOT_PARSED	1
+#define	NO_BODY				2
+#define	TRANSFER_ENCODING	3
+#define	CONTENT_LENGTH		4
+
 #endif
