+++ include/cgi_preprocessing.hpp
@@ -8,8 +8,8 @@
 # include "request.hpp"
 # include "parsing.hpp"
 # include "exec_request.hpp"
-
-# define BUF_SIZE 100
+# include "error.hpp"
+# include "../default_conf.hpp"
 
 class CGI_preprocessing
 {
@@ -22,14 +22,6 @@ public:
 	CGI_preprocessing(request &rq, conf_index &cf_i);
 	~CGI_preprocessing();
 	std::string	exec_CGI(void);
-
-	class	fork_error : public std::exception
-	{
-		virtual const char	*what(void) const throw()
-		{
-			return ("fork error!!");
-		}
-	};
 };
 
 #endif
